* commit '563c50768af9b4f8010ec82c449c231ad839f764':
  docs update
  Added explicit cast to Integer for NEW_RELIC_APP_ID

+++ vars/newrelic.groovy
@@ -28,7 +28,7 @@ def postDeployment(jobConfig) {
     }
 
     if (appId == 0 && jobConfig.NEW_RELIC_APP_ID && jobConfig.NEW_RELIC_APP_ID.containsKey(environment)) {
-        appId = jobConfig.NEW_RELIC_APP_ID[environment]
+        appId = jobConfig.NEW_RELIC_APP_ID[environment] as Integer
     }
 
     if (appId == 0 && jobConfig.NEW_RELIC_APP_NAME) {
